Major: Observation data input/storage modifications
Observation data in database now refers to the study directly and not to
an observation (data from previous schemas can not be loaded after these
changes). Affected tables are "observation_rounds" and "predefined" and
several functions, metrics and metric functions. Observation data input
has also been split to two servlets, one for getting validation data and
one for storing the data, in line with other types